+++ lib/cassette/rubycas/user_factory.rb
@@ -5,10 +5,11 @@ module Cassette
         attributes = session[:cas_extra_attributes]
         attributes = attributes.with_indifferent_access if attributes.respond_to?(:with_indifferent_access)
         Cassette::Authentication::User.new(login: session[:cas_user],
-                                           name: attributes.try(:[], :cn),
-                                           email: attributes.try(:[], :email),
-                                           authorities: attributes.try(:[], :authorities),
-                                           type: attributes.try(:[], :type).try(:downcase))
+                                           name: attributes.try(:delete, :cn),
+                                           email: attributes.try(:delete, :email),
+                                           authorities: attributes.try(:delete, :authorities),
+                                           type: attributes.try(:delete, :type).try(:downcase),
+                                           extra_attributes: attributes)
       end
     end
   end